feat: opened crew page 퍼블리싱
