--- public/constants/constants.js
@@ -1,15 +1,0 @@
-export const CLOSED_CREW_LIST = [
-  {
-    id: 0,
-    title: "이건 멋진 알고리즘 스터디",
-    category: "cate1",
-    content:
-      "이건개ㅐㅐㅐㅐㅐㅐㅐㅐㅐ 긴 설명이 될것입니당 구리릴이 링ㄹ이러ㅏㅇ러 ",
-  },
-  { id: 1, title: "title2", category: "cate1", content: "설명" },
-  { id: 2, title: "title3", category: "cate1", content: "설명" },
-  { id: 3, title: "title4", category: "cate1", content: "설명" },
-  { id: 4, title: "title5", category: "cate1", content: "설명" },
-  { id: 5, title: "title6", category: "cate1", content: "설명" },
-  { id: 6, title: "title7", category: "cate1", content: "설명" },
-];
